organized show editor related files
